# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM ИТ Школы РТК — REST API
  version: 0.2.0
  description: |
    API веб-клиента CRM и точка интеграции с корпоративными системами.

    Аутентификация: вход через Keycloak (GET /api/auth/login) создаёт серверную сессию BFF, браузер получает HttpOnly cookie SESSION; токены OIDC клиенту не передаются. Выход (POST /api/auth/logout) завершает сессию CRM и возвращает адрес завершения сессии Keycloak.

    Изменяющие запросы передают заголовок X-CSRF-TOKEN со значением token из GET /api/csrf. Команды передают Idempotency-Key; повтор с тем же ключом и телом возвращает прежний результат, с другим телом — 409 IDEMPOTENCY_CONFLICT. Устаревшая version даёт 409 VERSION_CONFLICT с currentVersion.

    Ошибки возвращаются в формате ApiError {code, message, requestId}. requestId совпадает с заголовком X-Request-Id и записью в журнале сервера; stack trace и секреты в ответ не попадают.
servers:
  - url: /
...
paths:
...
  /api/imports/inspect:
    post:
      tags: [Catalog]
      summary: Прочитать листы и заголовки файла импорта
      operationId: inspectCatalogImport
      x-implementation-stage: 4
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Workbook sheets and headers after signature and scanner checks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImportInspect'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/imports/preview:
    post:
      tags: [Catalog]
      summary: Предпросмотр импорта каталогов
      operationId: previewCatalogImport
      x-implementation-stage: 4
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, profile, sheet, mapping]
              properties:
                file:
                  type: string
                  format: binary
                profile:
                  $ref: '#/components/schemas/CatalogImportProfile'
                sheet:
                  type: string
                  minLength: 1
                  maxLength: 255
                mapping:
                  $ref: '#/components/schemas/CatalogImportMapping'
      responses:
        '202':
          description: Persisted terminal preview job and import plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImportPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/imports/{id}:
    get:
      tags: [Catalog]
      summary: Протокол импорта каталогов
      operationId: getCatalogImport
      x-implementation-stage: 4
      parameters:
        - $ref: '#/components/parameters/UuidId'
      responses:
        '200':
          description: Persisted import preview and row protocol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/imports/{id}/apply:
    post:
      tags: [Catalog]
      summary: Применить импорт каталогов
      operationId: applyCatalogImport
      x-implementation-stage: 4
      parameters:
        - $ref: '#/components/parameters/UuidId'
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogImportApply'
      responses:
        '202':
          description: Persisted terminal apply job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImportApplyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: Серверная сессия BFF, созданная входом через GET /api/auth/login. Swagger UI на том же origin использует сессию браузера.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Значение token из GET /api/csrf для текущей сессии. Без заголовка или с чужим значением запрос отклоняется с 403 FORBIDDEN.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    BadRequest:
      description: Validation or request format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: BFF session is absent or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated identity is not allowed to perform the action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Object does not exist or is outside the current server-side scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Version or idempotency conflict; VERSION_CONFLICT includes currentVersion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    PayloadTooLarge:
      description: Uploaded payload exceeds the configured limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalError:
      description: Unexpected server error; the stack trace is written only to the server log under the same requestId
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHENTICATED
            - FORBIDDEN
            - CRM_PROFILE_REQUIRED
            - CRM_PROFILE_PENDING
            - LAST_ACTIVE_ADMIN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_CONFLICT
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - RATE_LIMITED
            - INTERNAL_ERROR
            - DEPENDENCY_UNAVAILABLE
            - REPORT_NOT_READY
            - REPORT_ACCESS_CHANGED
            - REPORT_RESULT_UNAVAILABLE
            - REPORT_CAPACITY_EXCEEDED
            - SOURCE_NOT_CONFIGURED
            - SYNC_ALREADY_RUNNING
            - SYNC_CAPACITY_EXCEEDED
            - LMS_NOT_MAPPED
            - LMS_SYNC_FAILED
        message:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        currentVersion:
          type: integer
          minimum: 0
...
    CatalogImportProfile:
      type: string
      enum: [AGREEMENT, DIRECTION_PROGRAM]
      description: >-
        AGREEMENT loads the ten TZ catalog fields; DIRECTION_PROGRAM loads directions and programs by stable keys.
        Matching rules are described in docs/import.md.
    CatalogImportRowTarget:
      type: object
      description: >-
        Optional explicit record choice for one workbook row. An explicit id takes precedence over external keys
        and names in both preview and apply.
      properties:
        organizationId:
          type: [string, 'null']
          format: uuid
        managerProfileId:
          type: [string, 'null']
          format: uuid
          description: Active USER profile with a team; resolves a manager name shared by several profiles.
        interactionId:
          type: [string, 'null']
          format: uuid
        productAgreementId:
          type: [string, 'null']
          format: uuid
    CatalogImportMapping:
      type: object
      required: [columns]
      properties:
        columns:
          type: object
          description: >-
            Import field to workbook header. AGREEMENT requires organizationName, vendorName and productName and
            accepts contractNumber, licenseSigned, licenseExpiryYear, transferStatus, managerName, contactName,
            comment, organizationType, organizationExternalKey, vendorExternalKey, productExternalKey,
            agreementExternalKey, contactExternalKey, contactPosition, contactEmail, contactPhone and interactionId.
            Missing external keys are derived from normalized names. DIRECTION_PROGRAM requires
            directionExternalKey and directionName and accepts programExternalKey, programName and
            programDirectionRef.
          additionalProperties:
            type: string
            minLength: 1
            maxLength: 255
        rowTargets:
          type: object
          description: Workbook row number (2 or greater) to explicit record ids.
          additionalProperties:
            $ref: '#/components/schemas/CatalogImportRowTarget'
        transferStatuses:
          type: object
          description: >-
            Optional dictionary from a workbook transfer status to the stored value. Without a dictionary the
            normalized workbook value is stored as is.
          additionalProperties:
            type: string
            minLength: 1
            maxLength: 160
    CatalogImportSheet:
      type: object
      required: [name, headers]
      properties:
        name:
          type: string
        headers:
          type: array
          items:
            type: string
    CatalogImportInspect:
      type: object
      required: [sheets]
      properties:
        sheets:
          type: array
          items:
            $ref: '#/components/schemas/CatalogImportSheet'
    CatalogImportPreview:
      type: object
      required: [jobId, importId]
      properties:
        jobId:
          type: string
          format: uuid
        importId:
          type: string
          format: uuid
    CatalogImportRow:
      type: object
      required: [id, sheetName, rowNumber, status, fieldErrors, oldValues, newValues, applied]
      properties:
        id:
          type: string
          format: uuid
        sheetName:
          type: string
        rowNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [CREATE, UPDATE, UNCHANGED, CONFLICT, INVALID]
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        oldValues:
          type: object
          additionalProperties:
            type: string
        newValues:
          type: object
          additionalProperties:
            type: string
        applied:
          type: boolean
    CatalogImport:
      type: object
      required: [id, profile, status, version, rows, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        profile:
          $ref: '#/components/schemas/CatalogImportProfile'
        status:
          type: string
          enum: [PREVIEWED, APPLIED]
        version:
          type: integer
          minimum: 0
        rows:
          type: array
          items:
            $ref: '#/components/schemas/CatalogImportRow'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatalogImportApply:
      type: object
      required: [version, confirmedRowIds]
      properties:
        version:
          type: integer
          minimum: 0
        confirmedRowIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    CatalogImportApplyResult:
      type: object
      required: [jobId, importId]
      properties:
        jobId:
          type: string
          format: uuid
        importId:
          type: string
          format: uuid
...
security:
  - cookieAuth: []